Extracts status code checks to trait
